+++ examples/inference/src/bin/main.rs
@@ -1,0 +1,7 @@
+use loco_rs::cli;
+use inference::app::App;
+
+#[tokio::main]
+async fn main() -> eyre::Result<()> {
+    cli::main::<App>().await
+}