+++ examples/inference/tests/cli_tests.rs
@@ -1,0 +1,5 @@
+#[test]
+fn cli_tests() {
+    let t = trycmd::TestCases::new();
+    t.case("tests/cmd/*.trycmd");
+}