+++ src/boot.rs
@@ -304,7 +304,7 @@ pub fn run_app<H: Hooks>(mode: &StartMode, app_context: AppContext) -> Result<Bo
     match mode {
         StartMode::ServerOnly => {
             let app = H::routes().to_router(app_context.clone())?;
-            let router = H::after_router(app, &app_context)?;
+            let router = H::after_routes(app, &app_context)?;
             Ok(BootResult {
                 app_context,
                 router: Some(router),
@@ -314,7 +314,7 @@ pub fn run_app<H: Hooks>(mode: &StartMode, app_context: AppContext) -> Result<Bo
         StartMode::ServerAndWorker => {
             let processor = create_processor::<H>(&app_context)?;
             let app = H::routes().to_router(app_context.clone())?;
-            let router = H::after_router(app, &app_context)?;
+            let router = H::after_routes(app, &app_context)?;
             Ok(BootResult {
                 app_context,
                 router: Some(router),
@@ -367,7 +367,6 @@ pub fn list_endpoints<H: Hooks>() -> Vec<ListRoutes> {
 /// Initializes an [`EmailSender`] based on the mailer configuration settings
 /// ([`config::Mailer`]).
 fn create_mailer(config: &config::Mailer) -> Result<Option<EmailSender>> {
-    #[cfg(feature = "testing")]
     if config.stub {
         return Ok(Some(EmailSender::stub()));
     }