+++ CHANGELOG.md
@@ -2,14 +2,24 @@
 
 ## vNext
 
-* Added: `Hooks::before_run` with a default blank implementation. You can now code some custom loading of resources or other things before the app runs
-* Added: an LLM inference example, text generation in Rust, using an API (`examples/inference`)
+NEW FEATURES
+* `Hooks::before_run` with a default blank implementation. You can now code some custom loading of resources or other things before the app runs
+* an LLM inference example, text generation in Rust, using an API (`examples/inference`)
+* Loco starters version & create release script [https://github.com/loco-rs/loco/pull/110](https://github.com/loco-rs/loco/pull/110)
 
+DOCS:
+* Remove duplicated docs in auth section
+
+ENHANCEMENTS:
+* Remove unused libs: [https://github.com/loco-rs/loco/pull/106](https://github.com/loco-rs/loco/pull/106)
 
 ## 0.1.5
 
-* Added: `format:html` [https://github.com/loco-rs/loco/issues/74](https://github.com/loco-rs/loco/issues/74)
-* Added: Create a stateless html starter [https://github.com/loco-rs/loco/pull/100](https://github.com/loco-rs/loco/pull/100)
+NEW FEATURES
+* `format:html` [https://github.com/loco-rs/loco/issues/74](https://github.com/loco-rs/loco/issues/74)
+* Create a stateless HTML starter [https://github.com/loco-rs/loco/pull/100](https://github.com/loco-rs/loco/pull/100)
 * Added worker generator + adding a way to test workers [https://github.com/loco-rs/loco/pull/92](https://github.com/loco-rs/loco/pull/92)
-* Fix:  cargo cli to support fork prs [https://github.com/loco-rs/loco/pull/96](https://github.com/loco-rs/loco/pull/96)
+
+ENHANCEMENTS:
+* CI: allows cargo cli run on fork prs [https://github.com/loco-rs/loco/pull/96](https://github.com/loco-rs/loco/pull/96)