+++ CHANGELOG.md
@@ -6,6 +6,8 @@ NEW FEATURES
 * `Hooks::before_run` with a default blank implementation. You can now code some custom loading of resources or other things before the app runs
 * an LLM inference example, text generation in Rust, using an API (`examples/inference`)
 * Loco starters version & create release script [https://github.com/loco-rs/loco/pull/110](https://github.com/loco-rs/loco/pull/110)
+* Configure Cors middleware [https://github.com/loco-rs/loco/pull/114](https://github.com/loco-rs/loco/pull/114)
+* `Hooks::after_router` Invoke this function after the Loco routers have been constructed. This function enables you to configure custom Axum logics, such as layers, that are compatible with Axum. [https://github.com/loco-rs/loco/pull/114](https://github.com/loco-rs/loco/pull/114)
 
 DOCS:
 * Remove duplicated docs in auth section